video:tegra:hdcp: Change command for TZ operation
The command id for handling cmac operation is 0xB in TZ.
This patch fixes the command id from 0xA to 0xB so that the
correct command gets invoked. Essentially the command is for
obtaining the cmac of the srm version inside the HDCP TA. Once
the cmac is returned to kernel, it is passed down to the HDCP
ucode for sanctity checks.

Bug 200292655

Change-Id: Id2cf66404cef116bca0cae0c99e4d377e82e3924
Signed-off-by: Pranami Bhattacharya <pranamib@nvidia.com>
Reviewed-on: http://git-master/r/1484129
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Vinayak Pane <vpane@nvidia.com>