usb: cdc-acm: Add DISABLE_AUTOSUSPEND quirk
This change adds a new DISABLE_AUTOSUSPEND quirk. If the quirk is set,
we signal to the usb system to never attempt to autosuspend a device.

The ST Link dongle does not support autosuspend at all. Despite having
reset on resume, they go in a bad state if we try. Hence this
change adds the ST Link dongle to the quirk list to disable
autosuspend.

Fixes the issues with the dongle not responding on boot.

Bug 200258732

Change-Id: I1570f0e05c6e7623d0247ca158920c621e064420
Signed-off-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1529223
(cherry picked from commit 1b2f9269062d80a54cc9c15fffaeddfc599ae6ff)
Reviewed-on: https://git-master.nvidia.com/r/1590246
GVS: Gerrit_Virtual_Submit
Reviewed-by: Vinayak Pane <vpane@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>