video: tegra: dc: fix multiple HPD in HDCP
This change clears the g_fallback flag only when there
is real HPD. HDCP state machine uses g_fallback to decide
if it has to authenticate 2.2 or 1.4 on this. So it skips
multiple HPD to be notified to userspace. Earlier to this
change, when HDCP authentication falls from 2.2 to 1.4,
it used to send hpd unplug and plug event.

bug 1744655

Change-Id: Idd778cfab918e95b6c8c72ad2a684787e673bd9a
Signed-off-by: Prafull Suryawanshi <prafulls@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1595883
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mitch Luban <mluban@nvidia.com>
Tested-by: Mitch Luban <mluban@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>