hdcp: hdmi: skip hdcp operation when vedid is set
Even When vedid is set, hdcp ioctl calls set the
HDCP to be turned on for HDCP auth to take place.
HDCP authentication try to access i2c lines and
sor registers even when display not connected.
Also it can keep worker thread rescheduling and
causing access to sor registers result in warnings.
To fix it, check if vedid is set and then only
proceed further to carry out hdcp authentication.

Bug 200370553

Change-Id: I7af768a573403d8e91c64abdad1970a200de3615
Signed-off-by: Sujeet Baranwal <sbaranwal@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1642440
Reviewed-by: Pranami Bhattacharya <pranamib@nvidia.com>
Reviewed-by: Vinayak Pane <vpane@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>