hdmivrr: fix gsync mode signalling to monitor
As per the GSYNC requirements, we are expected to send 2 extra lines as
part of vbp to indicate to the monitor that we are running VRR mode.
This is done as part of _tegra_hdmivrr_activate.

However, _tegra_dc_set_mode doesn't set mode_dirty flag if it isn't YUV420.
tegra_dc_update_mode doesn't update the HW registers if mode_dirty flag
isn't set.

So explicitly set mode_dirty flag to true before invoking
tegra_dc_update_mode. With this fix, when tegradc is in VRR mode, the
monitor will switch from normal mode to G-Sync mode.

Bug 2061491

Change-Id: I1489d3f8c0db2586cbafebb7a7db5f94fd8f267a
Signed-off-by: Ishwarya Balaji Gururajan <igururajan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1662597
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>