hid: jarvis: Move setup_timer to avoid k-crash
The setup_timer() function internally calls init_timer() which
_could_ move the pcm timer's base to a new CPU base and lead to an
inconsistency in the timer's local copy of the base w.r.t the
original base.
Ideally, we need to make sure that the timer is not pending or
active when init_timer() is called.

This change moves the setup_timer() call to the snd_pcm_open()
function so that the timer's base doesn't change as long as the
stream is open (a.k.a timer is being used).

Bug 200352102

Change-Id: I5e5ad29c729a341496ff8e91b525a53c9db84a02
Reviewed-on: https://git-master.nvidia.com/r/1670047
GVS: Gerrit_Virtual_Submit
Tested-by: Abhiroop Kaginalkar <akaginalkar@nvidia.com>
Reviewed-by: Martin Gao <marting@nvidia.com>
Reviewed-by: Vinayak Pane <vpane@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>