driver: therm_fan_est: fix across 2 tripponit issue
When we have the stress test with Start/Stop high CPU load,
the Fan will switch between different speed according to the temp.
After a period of time, there is a chance that the Fan get stuck in
high seepd even the temp is drop down.

therm_fan_est support the hysteresis, we will change the Fan speed
in different throttle points, but if the temp drop down quickly and
across 2 hysteresis trippoints in one time, we will fail to handle it
in current logic. Result in missing to update the Fan speed.

Take across 2 hsystereis tirppoints case into account for the thermal
zone udpate will fix this issue.

Bug 2256802

Change-Id: I5472d94f86a458c2ed85d389e2a75754c232e4db
Signed-off-by: Daniel Fu <danifu@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1792091
GVS: Gerrit_Virtual_Submit
Reviewed-by: Vinayak Pane <vpane@nvidia.com>